- applied a first patch from Gunther Nikl to try to get the OS3 version of amissl_vXXX.library running with baserel+multibase.

+++ libcmt/file.c
@@ -206,7 +206,7 @@ int close(UNUSED int fd)
 	return(-1);
 }
 
-int lseek(UNUSED int fd, UNUSED long offset, UNUSED int mode)
+off_t lseek(UNUSED int fd, UNUSED long offset, UNUSED int mode)
 {
 	SetAmiSSLerrno(EINVAL);
